Feature/security updated

+++ src/main/java/net/weg/taskmanager/repository/OptionRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 
-public interface OptionRepository extends JpaRepository<Option, Integer> {
+public interface OptionRepository extends JpaRepository<Option, Long> {
 }
